[feature] add exception test

+++ DeviceDriver/DeviceDriver.cpp
@@ -10,7 +10,7 @@ void DeviceDriver::read5times(long address, int readResult[5])
 {
 	for (int i = 0; i < 5; ++i)
 	{
-		readResult[i] = m_hardware->read(address);
+		readResult[i] = (int)m_hardware->read(address);
 		Sleep(200);
 	}
 }
@@ -31,7 +31,7 @@ int DeviceDriver::read(long address)
     int readResult[5] = { 0, };
     read5times(address, readResult);
     checkReadValidation(readResult);
-
+	
     return (int)(readResult[0]);
 }
